# C# write benchmark improvements