- target netcoreapp3.1 (the newest LTS of .NET core)
- add a few more useful write benchmarks
- use `BinaryPrimitives.WriteUInt64LittleEndian` for writing fixed64 values as starting from .NET Core 3.x, this is much faster than the current workaround.

Changes are mostly taken from https://github.com/JamesNK/protobuf/commit/337780c0e098185d4df76d498c8f435e105f8a56 (as recommended here https://github.com/protocolbuffers/protobuf/pull/7576#issuecomment-647069227 for more context). Thanks @JamesNK  for the improvements!