@jtattermusch 2020-06-29
Comparison for using `BinaryPrimitives.WriteUInt64LittleEndian` on .NET Core 3.1
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18362.900 (1903/May2019Update/19H1)
Intel Core i7-7600U CPU 2.80GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.1.301
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT

BEFORE:

|                    Method | BytesToWrite |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| WriteFixed64_WriteContext |        10080 | 14.97 us | 0.156 us | 0.138 us |     - |     - |     - |         - |

AFTER:

|                    Method | BytesToWrite |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |------------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| WriteFixed64_WriteContext |        10080 | 3.097 us | 0.0603 us | 0.1071 us |     - |     - |     - |         - |
```