@jtattermusch 2020-06-29
the CI doesn't have a new-enough version of the SDK (which is something that would need to be fixed first).
```
dotnet restore csharp/src/Google.Protobuf.sln
/root/.dotnet/sdk/3.0.100/Sdks/Microsoft.NET.Sdk/targets/Microsoft.NET.TargetFrameworkInference.targets(127,5): error NETSDK1045: The current .NET SDK does not support targeting .NET Core 3.1.  Either target .NET Core 3.0 or lower, or use a version of the .NET SDK that supports .NET Core 3.1. [/tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Benchmarks/Google.Protobuf.Benchmarks.csproj]
```