@JamesNK 2020-06-30
Opps, I saw 3.x in global.json and assumed it was 3.1.

Build should update to 3.1. It is a LTS release. 3.0 is no longer supported.